requirements.txt
Django==3.1.4
djangorestframework==3.12.2
mysqlclient==2.0.2
django-filter==2.4.0
